Rename Ability to BaseAbility, Create interface Ability
